refactor: break massive file usage.ts into multiple files for cleaner code.
